backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2022.12.7
click==8.1.8
distlib==0.3.6
exceptiongroup==1.3.0
fastapi==0.115.14
filelock==3.9.0
h11==0.16.0
idna==3.10
pipenv==2023.2.18
platformdirs==3.0.0
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
python-multipart==0.0.20
sniffio==1.3.1
SQLAlchemy==2.0.42
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
virtualenv==20.20.0
virtualenv-clone==0.5.7

# --- NEW for auth ---
passlib[bcrypt]==1.7.4
PyJWT==2.9.0

